Updated recovery scripts for recovery 19.07.8
The recovery scripts have been rewritten for the new recovery 19.07.8.
Besides the capability to mount/unmount the main firmware image, it is now
possible to reset the main to defaults.
To facilitate the unlocking of the MTD partitions a new script was also
included as the interface to the mtd-rw kernel module.

Signed-off-by: Bruno Pena <brunompena@gmail.com>